[AIP-133] Document all AIP-133 rules.
Also cleans up AIP-131 rules ever so slightly.
