+++ docs/rules/core/0131-response-message-name.md
@@ -7,8 +7,8 @@ rule:
 
 # Upper snake case values
 
-This rule enforces that all `Get*` RPCs have a response message of the
-resource, as mandated in [AIP-131][].
+This rule enforces that all `Get` RPCs have a response message of the resource,
+as mandated in [AIP-131][].
 
 ## Details
