+++ docs/rules/core/0133-http-uri-parent.md
@@ -1,0 +1,65 @@
+---
+rule:
+  aip: 133
+  name: [core, '0133', http-uri-parent]
+  summary: Create methods must map the parent field to the URI.
+---
+
+# Get methods: Request message
+
+This rule enforces that all `Create` RPCs map the `name` field to the HTTP URI,
+as mandated in [AIP-133][].
+
+## Details
+
+This rule looks at any message matching beginning with `Create`, and complains
+if the `parent` variable is not included in the URI. It _does_ check additional
+bindings if they are present.
+
+## Examples
+
+**Incorrect** code for this rule:
+
+```proto
+// Incorrect.
+rpc CreateBook(CreateBookRequest) returns (Book) {
+  option (google.api.http) = {
+    post: "/v1/publishers/*/books"  // The `parent` field should be extracted.
+    body: "book"
+  };
+}
+```
+
+**Correct** code for this rule:
+
+```proto
+// Correct.
+rpc CreateBook(CreateBookRequest) returns (Book) {
+  option (google.api.http) = {
+    post: "/v1/{parent=publishers/*}/books"
+    body: "book"
+  };
+}
+```
+
+## Disabling
+
+If you need to violate this rule, use a leading comment above the method.
+Remember to also include an [aip.dev/not-precedent][] comment explaining why.
+
+```proto
+// (-- api-linter: core::0133::http-uri-parent=disabled
+//     aip.dev/not-precedent: We need to do this because reasons. --)
+rpc CreateBook(CreateBookRequest) returns (Book) {
+  option (google.api.http) = {
+    post: "/v1/publishers/*/books"
+    body: "book"
+  };
+}
+```
+
+If you need to violate this rule for an entire file, place the comment at the
+top of the file.
+
+[aip-133]: https://aip.dev/133
+[aip.dev/not-precedent]: https://aip.dev/not-precedent
