+++ docs/rules/core/0133-request-message-name.md
@@ -1,14 +1,14 @@
 ---
 rule:
-  aip: 131
-  name: [core, '0131', request-message-name]
-  summary: Get methods must have standardized request message names.
+  aip: 133
+  name: [core, '0133', request-message-name]
+  summary: Create methods must have standardized request message names.
 ---
 
-# Get methods: Request message
+# Create methods: Request message
 
 This rule enforces that all `Get` RPCs have a request message name of
-`Get*Request`, as mandated in [AIP-131][].
+`Get*Request`, as mandated in [AIP-133][].
 
 ## Details
 
@@ -22,10 +22,12 @@ with the suffix `Request` appended.
 
 ```proto
 // Incorrect.
-rpc GetBook(GetBookReq) returns (Book) {  // Should be `GetBookRequest`.
+// Should be `CreateBookRequest`.
+rpc CreateBook(CreateBookReq) returns (Book) {
   option (google.api.http) = {
-    get: "/v1/{name=publishers/*/books/*}"
-  }
+    post: "/v1/{parent=publishers/*}/books"
+    body: "book"
+  };
 }
 ```
 
@@ -33,10 +35,11 @@ rpc GetBook(GetBookReq) returns (Book) {  // Should be `GetBookRequest`.
 
 ```proto
 // Correct.
-rpc GetBook(GetBookRequest) returns (Book) {
+rpc CreateBook(CreateBookRequest) returns (Book) {
   option (google.api.http) = {
-    get: "/v1/{name=publishers/*/books/*}"
-  }
+    post: "/v1/{parent=publishers/*}/books"
+    body: "book"
+  };
 }
 ```
 
@@ -46,17 +49,18 @@ If you need to violate this rule, use a leading comment above the method.
 Remember to also include an [aip.dev/not-precedent][] comment explaining why.
 
 ```proto
-// (-- api-linter: core::0131::request-message-name=disabled
+// (-- api-linter: core::0133::request-message-name=disabled
 //     aip.dev/not-precedent: We need to do this because reasons. --)
-rpc GetBook(GetBookReq) returns (Book) {
+rpc CreateBook(CreateBookReq) returns (Book) {
   option (google.api.http) = {
-    get: "/v1/{name=publishers/*/books/*}"
-  }
+    post: "/v1/{parent=publishers/*}/books"
+    body: "book"
+  };
 }
 ```
 
 If you need to violate this rule for an entire file, place the comment at the
 top of the file.
 
-[aip-131]: https://aip.dev/131
+[aip-133]: https://aip.dev/133
 [aip.dev/not-precedent]: https://aip.dev/not-precedent
