+++ rules/aip0133/request_unknown_fields_test.go
@@ -22,124 +22,6 @@ import (
 	"github.com/jhump/protoreflect/desc/builder"
 )
 
-type field struct {
-	fieldName string
-	fieldType *builder.FieldType
-}
-
-func TestParentField(t *testing.T) {
-	// Set up the testing permutations.
-	tests := []struct {
-		testName     string
-		messageName  string
-		messageField *field
-		problems     testutils.Problems
-		problemDesc  func(m *desc.MessageDescriptor) desc.Descriptor
-	}{
-		{
-			"Valid",
-			"CreateBookRequest",
-			&field{"parent", builder.FieldTypeString()},
-			testutils.Problems{},
-			nil},
-		{
-			"MissingField",
-			"CreateBookRequest",
-			nil,
-			testutils.Problems{{Message: "parent"}},
-			nil,
-		},
-		{
-			"InvalidType",
-			"CreateBookRequest",
-			&field{"parent", builder.FieldTypeDouble()},
-			testutils.Problems{{Message: "string"}},
-			func(m *desc.MessageDescriptor) desc.Descriptor {
-				return m.FindFieldByName("parent")
-			},
-		},
-	}
-
-	// Run each test individually.
-	for _, test := range tests {
-		t.Run(test.testName, func(t *testing.T) {
-			// Create an appropriate message descriptor.
-			messageBuilder := builder.NewMessage(test.messageName)
-
-			if test.messageField != nil {
-				messageBuilder.AddField(
-					builder.NewField(test.messageField.fieldName, test.messageField.fieldType),
-				)
-			}
-
-			message, err := messageBuilder.Build()
-			if err != nil {
-				t.Fatalf("Could not build %s message.", test.messageName)
-			}
-
-			// Determine the descriptor that a failing test will attach to.
-			var problemDesc desc.Descriptor = message
-			if test.problemDesc != nil {
-				problemDesc = test.problemDesc(message)
-			}
-
-			// Run the lint rule, and establish that it returns the correct problems.
-			problems := parentField.Lint(message.GetFile())
-			if diff := test.problems.SetDescriptor(problemDesc).Diff(problems); diff != "" {
-				t.Errorf(diff)
-			}
-		})
-	}
-}
-
-func TestResourceField(t *testing.T) {
-	// Set up the testing permutations.
-	tests := []struct {
-		testName     string
-		messageName  string
-		messageField *field
-		problems     testutils.Problems
-	}{
-		{
-			"Valid",
-			"CreateBookRequest",
-			&field{"book", builder.FieldTypeMessage(builder.NewMessage("Book"))},
-			testutils.Problems{},
-		},
-		{
-			"MissingField",
-			"CreateBookRequest",
-			nil,
-			testutils.Problems{{Message: "Message \"CreateBookRequest\" has no \"Book\" type field"}},
-		},
-	}
-
-	// Run each test individually.
-	for _, test := range tests {
-		t.Run(test.testName, func(t *testing.T) {
-			// Create an appropriate message descriptor.
-			messageBuilder := builder.NewMessage(test.messageName)
-
-			if test.messageField != nil {
-				messageBuilder.AddField(
-					builder.NewField(test.messageField.fieldName, test.messageField.fieldType),
-				)
-			}
-
-			message, err := messageBuilder.Build()
-			if err != nil {
-				t.Fatalf("Could not build %s message.", test.messageName)
-			}
-
-			// Run the lint rule, and establish that it returns the correct problems.
-			problems := resourceField.Lint(message.GetFile())
-			if diff := test.problems.SetDescriptor(message).Diff(problems); diff != "" {
-				t.Errorf(diff)
-			}
-		})
-	}
-}
-
 func TestUnknownFields(t *testing.T) {
 	// Set up the testing permutations.
 	tests := []struct {
